@@ -1,7 +1,6 @@
 plugins {
     id 'groovy'
-    id 'java-gradle-plugin'
-    id 'fabric-loom' version '1.2-SNAPSHOT'
+    id 'fabric-loom' version '1.6-SNAPSHOT'
     id 'maven-publish'
     id "com.modrinth.minotaur" version "2.+"
     id 'com.matthewprenger.cursegradle' version "1.4.0"
@@ -32,7 +31,7 @@ allprojects {
     version = full_version
     archivesBaseName = "${project.mod_id}"
 
-    sourceCompatibility = targetCompatibility = JavaVersion.VERSION_17
+    sourceCompatibility = targetCompatibility = JavaVersion.VERSION_21
 
     // Declare dependencies
     dependencies {
@@ -115,7 +114,6 @@ allprojects {
         // Loom adds the essential maven repositories to download Minecraft and libraries from automatically.
         // See https://docs.gradle.org/current/userguide/declaring_repositories.html
         // for more information about repositories.
-        maven { url 'https://oss.sonatype.org/content/repositories/snapshots' }
         maven { url 'https://maven.nucleoid.xyz' }
         // Used by FabricPlaceholderAPI
 
